Fixing relative paths for moved tests
